### 2023-10-30 #5997 @Kevin-Mattheus-Moerman
@WillHannon-MCB thanks for this submission. I am the AEiC for this track. I have just triggered a scope review by the editorial board for this submission. This is because I need some help to see if this work is in scope, both in terms of its significance/size (~2781) and also functionality (online based visualisation tool). Scope reviews usually take about 2 weeks to complete.

@WillHannon-MCB it would be helpful if you could indicate what portion of the JavaScript has been your new original development, and what portion has for instance been autogenerated or mostly copied, e.g. to form somewhat standard supporting GUI structures.
