### 2023-10-30 #5997 @WillHannon-MCB
Hi @Kevin-Mattheus-Moerman,

All of the code is new/original development. The most boilerplate is probably in the [ui.js](https://github.com/dms-viz/dms-viz.github.io/blob/submit-joss-manuscript/v0/js/ui.js) file.

I should also mention that there is a small python utility for formatting user data that goes along with the web-based visualization tool [here](https://github.com/dms-viz/configure_dms_viz).

Thanks for your help and let me know if there is anything else you need.
Will
