### 2023-11-01 d644998
Update the tooltip when the data changes
