+++ assets/v0.ef9f70f3.js
@@ -4898,10 +4898,10 @@ vClipCenter = -( modelViewMatrix * vec4( clipCenter, 1.0 ) ).xyz;
 `)!==-1&&(e=e.replace(/\\\n/g,""));var r=e.split(`
 `),a="",s="",o="",c=[],A=typeof"".trimLeft=="function";for(t=0,n=r.length;t<n;t++)if(a=r[t],(a=A?a.trimLeft():a.trim()).length!==0&&(s=a.charAt(0))!=="#"){if(s==="v"){if((o=a.charAt(1))===" "&&(c=this.regexp.vertex_pattern.exec(a))!==null)i.vertices.push(parseFloat(c[1]),parseFloat(c[2]),parseFloat(c[3]));else if(o==="n"&&(c=this.regexp.normal_pattern.exec(a))!==null)i.normals.push(parseFloat(c[1]),parseFloat(c[2]),parseFloat(c[3]));else if(o!=="t"||this.regexp.uv_pattern.exec(a)===null)throw new Error("Unexpected vertex/normal/uv line: '"+a+"'")}else if(s==="f"){if((c=this.regexp.face_vertex_uv_normal.exec(a))!==null)i.addFace(c[1],c[4],c[7],c[10],c[3],c[6],c[9],c[12]);else if(this.regexp.face_vertex_uv.exec(a)===null)if((c=this.regexp.face_vertex_normal.exec(a))!==null)i.addFace(c[1],c[3],c[5],c[7],c[2],c[4],c[6],c[8]);else{if((c=this.regexp.face_vertex.exec(a))===null)throw new Error("Unexpected face line: '"+a+"'");i.addFace(c[1],c[2],c[3],c[4])}}else if(s==="l"){var l=a.substring(1).trim().split(" "),u=[],d=[];if(a.indexOf("/")===-1)u=l;else for(var p=0,h=l.length;p<h;p++){var g=l[p].split("/");g[0]!==""&&u.push(g[0]),g[1]!==""&&d.push(g[1])}i.addLineGeometry(u,d)}else if((c=this.regexp.object_pattern.exec(a))!==null){var b=c[0].substr(1).trim();i.startObject(b)}else if(!this.regexp.material_use_pattern.test(a)){if(!this.regexp.material_library_pattern.test(a)){if(this.regexp.smoothing_pattern.exec(a)===null){if(a==="\0")continue;throw new Error("Unexpected line: '"+a+"'")}}}}var y=[];for(t=0,n=i.objects.length;t<n;t++){var v=i.objects[t].geometry;if(v.vertices.length!==0){var _=new Ct;_.setAttribute("position",new Vt(new Float32Array(v.vertices),3)),v.normals.length>0?_.setAttribute("normal",new Vt(new Float32Array(v.normals),3)):_.computeVertexNormals(),y.push(_)}}return y}};var Pde=function(e){function t(){e.apply(this,arguments)}e&&(t.__proto__=e),t.prototype=Object.create(e&&e.prototype),t.prototype.constructor=t;var n={type:{configurable:!0}};return n.type.get=function(){return"obj"},t.prototype.getLoader=function(){return new ux},Object.defineProperties(t.prototype,n),t}(rL);Jt.add("obj",Pde);var dx=function(){this.propertyNameMapping={}};dx.prototype={constructor:dx,setPropertyNameMapping:function(e){this.propertyNameMapping=e},bin2str:function(e){for(var t=new Uint8Array(e),n="",i=0;i<e.byteLength;i++)n+=String.fromCharCode(t[i]);return n},isASCII:function(e){return this.parseHeader(this.bin2str(e)).format==="ascii"},parse:function(e){return e instanceof 
[…]
