### 2023-11-03 d8da182
Set filters to a default value
