### 2023-11-06 7625cdc
Fix issue with default filters overwritting URL
