### 2023-12-04 #5997 @Kevin-Mattheus-Moerman
@WillHannon-MCB apologies for the delay in processing the scope review. We have now concluded the scope review and determined this work may be in scope for JOSS. I will now search for a handling editor.
