### 2023-12-08 #5997 @afrubin
@csoneson This is my area and I'm happy to review the work's scientific content and usability. However, I'm not a JavaScript developer so I can't really comment on the code style and quality.
