### 2023-12-08 #6129 @editorialbot
[REVIEW]: dms-viz: Structure-informed visualizations for deep mutational scanning and other mutation-based datasets
**Submitting author:** @WillHannon-MCB (<a href="http://orcid.org/0000-0002-6014-4749">William Hannon</a>)
**Repository:** https://github.com/dms-viz/dms-viz.github.io
**Branch with paper.md** (empty if default branch): submit-joss-manuscript
**Version:** v0.8.5
