### 2024-01-25 dms-viz/dms-viz.github.io#58 @jkanche
Review for JOSS
JOSS: https://github.com/openjournals/joss-reviews/issues/6129

Overall, dms-viz appears to be a powerful tool for exploring complex 3D structures and mutational data. The effort invested in tutorials and user support is awesome. However, enhancing documentation for developer workflows would significantly improve the tool's usability.

Usability Recommendations:

- [x] #59
- [x] #80

- [x] #60
- [x] #81
  - Seems like when i brush or select the red dots; these are coded as white in the 3D visualization; which can be hard to notice
- [x] #61
- [x] #62

- [x] #63
- [x] #64
  - I notice there's a **click for description** in the bottom that contains some of this information
- [x] Fix typo's in the documentation

Code repositories:
- configure-dms-viz:
  - [x] Consider providing more information to the README on how to install the package and quick usage
  - [x] Improve package metadata on [pypi](https://pypi.org/project/configure-dms-viz/) for better discoverability
  - [x] Include information on testing and building documentation in the contribution markdown
  - [x] Consider using sphinx or other documentation tools to document this repository independent from the end user tutorial

- dms-viz: The repository seems to be well documented.
  - [x] #65
  - [x] #66
  - [x] #67
  - [x] #68
  - [x] #82

Improve the README on the organization page by providing more context about the application. Consider adding a screenshot of the application along with a link for easy access.
