### 2024-01-26 dms-viz/dms-viz.github.io#58 @WillHannon-MCB
Thank you for taking the time to review `dms-viz` and for your helpful comments! I'll start addressing these now. I'll address most of these separately as `issues` and tag them here when they're closed. Does that work for you?

Thanks again!
