### 2024-02-02 #6129 @afrubin
I'm looking forward to playing around with the tool some more and uploading some other DMS datasets, but I have some initial comments on the paper itself that I wanted to leave here since they aren't directly related to the software (and therefore less suited to becoming issues on that repo).

The introduction skips over a lot of basic information and launches right into some very specific applications from your lab. It seems like an extra sentence or two about the data and some key citations (like [Fowler & Fields 2014](https://pubmed.ncbi.nlm.nih.gov/25075907/) and some other more recent reviews) would be worth including. You might also want to cite the recent overview ["An Atlas of Variant Effects to understand the genome at nucleotide resolution"](https://pubmed.ncbi.nlm.nih.gov/37394429/) to show the expected expansion of the field.

There are other visualization tools like [ProtVar](https://www.embl.org/news/science/protvar-launch/) that could be mentioned.

The examples are currently all viral datasets, while most of the DMS data being generated is from human genes. This makes sense given the authors' research focus, but I recommend adding some more examples from diverse organisms generated by other labs to back up the claim that this is a general tool. There's a lot of data in [MaveDB](https://www.mavedb.org) and I'll share my results as I try to visualize some of my favorite examples.
