### 2024-02-06 dms-viz/dms-viz.github.io#69 @afrubin
Color options
When first exploring the example dataset, I had a bit of trouble understanding that there were multiple "tracks" and couldn't figure out why only the residues with high blue scores were showing up as blue.

It's also very hard to see what's changing with the default white space fill on white protein color scheme.

I think it would help first time users if the Chart Options section on the sidebar was expanded by default, making it more obvious that there are two datasets to consider each with its own color.

I'd also suggest choosing contrasting colors as the default values for Protein, Peripheral, and Ligand Color. This will make the selected residues pop out more even when they are white when colored by score.

This issue is part of https://github.com/openjournals/joss-reviews/issues/6129
