### 2024-02-06 dms-viz/dms-viz.github.io#70 @afrubin
Filter options is blank in the example dataset?
I wasn't able to see any content under the Filter Options for the example dataset, regardless of whether I had some residues selected or not. It's possible that I wasn't using the site correctly. Did I need to select something else first?

I'm using Firefox 122.0 (64-bit) installed via Snap for Ubuntu

This issue is part of https://github.com/openjournals/joss-reviews/issues/6129
