### 2024-02-06 dms-viz/dms-viz.github.io#71 @afrubin
Select All Sites should be a button
There's a checkbox called "Select All Sites" under Chart Options, but it seems like this should be a button instead. This is because it's a single operation the user would like to perform rather than a setting. Also, the box doesn't seem to check itself when all residues are selected manually (not to say that this would be better).

This issue is part of https://github.com/openjournals/joss-reviews/issues/6129
