### 2024-02-06 dms-viz/dms-viz.github.io#72 @afrubin
Reset structure doesn't work?
According the the help button (?) on the site, I should be able to reset the structure by pressing R on the keyboard. However, this doesn't work for my browser.

I'm using Firefox 122.0 (64-bit) installed via Snap for Ubuntu

This issue is part of https://github.com/openjournals/joss-reviews/issues/6129
