### 2024-02-06 dms-viz/dms-viz.github.io#73 @afrubin
Add median score
Median score at a position seems like a very sensible summary statistic that is currently missing from the selector under "Chart Options".

This issue is part of https://github.com/openjournals/joss-reviews/issues/6129
