### 2024-02-06 dms-viz/dms-viz.github.io#74 @afrubin
Add default filters
It would be nice to see some sensible default filtering options. Some examples that could be exposed for every datasets include:
* Wildtype amino acids
* Completeness at that position (i.e. number of possible amino acid substitutions observed)
* Score range

I'm sure there are other examples of default filters that are broadly applicable that could be added as well, but including some of these would help users explore their data and get more value out of the tool.

This issue is part of https://github.com/openjournals/joss-reviews/issues/6129
