### 2024-02-07 95f51f6
Add median as summary statistic
