### 2024-02-07 dms-viz/dms-viz.github.io#72 @WillHannon-MCB
Two issues could cause this, both my fault:

1. The instructions should say `r` and not `R` (as in `r` + `shift`).
2. I should also clarify in the instructions that by `reset`, I mean that if you've either _zoomed_ into the structure or _moved_ the structure within the window, pressing `r` will cause the structure to center in the window and reset to the default zoom. It will not reset the _orientation_ of the structure.

I'll update the instructions accordingly. If you still have issues, I assume it's a browser-specific error and I'll address that accordingly.
