### 2024-02-07 16b8a6f
Expand chart options by default
