### 2024-02-07 dms-viz/dms-viz.github.io#70 @WillHannon-MCB
There are no filters in the default dataset. But I will address this by either adding default filters or by hiding the filters if there aren't any. I'm addressing these in other issues, so I'm closing this one.
