### 2024-02-07 716e64b
Change the default protein colors to contrast
