+++ v0/js/tool.js
@@ -605,9 +605,9 @@ export class Tool {
       },
       backgroundRepresentation: { abbrev: "br", default: "rope", json: false },
       ligandRepresentation: { abbrev: "lr", default: "spacefill", json: false },
-      proteinColor: { nabbrevame: "pc", default: "#d3d3d3", json: false },
-      backgroundColor: { abbrev: "bc", default: "#d3d3d3", json: false },
-      ligandColor: { abbrev: "lc", default: "#d3d3d3", json: false },
+      proteinColor: { abbrev: "pc", default: "#363636", json: false },
+      backgroundColor: { abbrev: "bc", default: "#363636", json: false },
+      ligandColor: { abbrev: "lc", default: "#363636", json: false },
       ligandElement: { abbrev: "le", default: false, json: false },
       proteinElement: { abbrev: "pce", default: false, json: false },
       proteinOpacity: { abbrev: "po", default: "1", json: false },
