### 2024-02-07 dms-viz/dms-viz.github.io#69 @WillHannon-MCB
1. I made it so that the chart options are open by default.
2. I made the default protein, peripheral, and ligand colors darker so that they'd contrast with the selection.
