### 2024-02-08 2f54e70
Dynamically insert the version
