+++ vite.config.js
@@ -1,4 +1,7 @@
-export default {
+import { defineConfig } from "vite";
+import { version } from "./package.json";
+
+export default defineConfig({
   build: {
     rollupOptions: {
       input: {
@@ -7,4 +10,8 @@ export default {
       },
     },
   },
-};
+  define: {
+    // Pass the app version from package.json to the app
+    __APP_VERSION__: JSON.stringify(version),
+  },
+});
