### 2024-02-12 6463586
Take into account protein insersion codes
