### 2024-02-13 76288a6
Format selected sites with insertion codes
