### 2024-03-12 911f695
Add button to change background color of protein
