### 2024-03-12 e201f34
Remove help icon highlighting
