### 2024-03-13 68d7e54
Remove redundant event listener
