### 2024-03-13 dms-viz/dms-viz.github.io#74 @WillHannon-MCB
I've added a default mutation count filter. I'll think about other useful default filters.

I'm not planning to add a filter for the 'wildtype' amino acid. I've gotten feedback from some users that they're using data that doesn't have a meaningful value for the 'wildtype' amino acid. In the future, I might make specifying a value for wildtype optional.
