### 2024-03-14 4210226
Improve the timing of loading events
