### 2024-04-05 58fae41
Collapse help message by default
