### 2024-04-08 5ca4824
Add autoview on selection
