### 2024-04-09 869b2f7
Refactor main CSS components
