### 2024-04-10 f586bc7
Complete overhaul of CSS and UI
