+++ v0/js/ui.js
@@ -3,48 +3,67 @@ import "tippy.js/dist/tippy.css";
 
 export class UI {
   constructor() {
-    // UI Elements
-    this.infoButton = document.getElementById("dataset-info-button");
-    this.sidebar = document.getElementById("sidebar");
-    this.navbar = document.getElementById("nav-bar");
+    /* Main UI elements */
+    this.sidebar = document.getElementById("app-sidebar");
+    this.navbar = document.getElementById("app-navbar");
+    this.main = document.getElementById("app-main");
+
+    /* Elements responsible for toggling the sidebar */
     this.toggle = document.getElementById("sidebar-toggle");
-    this.headerpd = document.getElementById("header");
-    this.mainpd = document.getElementById("main");
+
+    /* Elements responsible for toggling the download buttons */
     this.localFile = document.getElementById("local-file");
     this.remoteFile = document.getElementById("remote-file");
-    this.accordionParent = document.getElementById("sidebar");
+
+    /* Register event listeners */
     this.registerEventListeners();
     this.registerTooltips();
   }
 
-  // UI functions to attach to event listeners
+  /* Method to toggle sidebar classes */
   toggleSidebar() {
-    return new Promise((resolve) => {
-      this.sidebar.classList.toggle("sidebar--collapsed");
-      this.navbar.classList.toggle("sidebar-collapsed");
-      this.mainpd.classList.toggle("sidebar-collapsed");
-      this.toggle.classList.toggle("bx-x");
-      this.headerpd.classList.toggle("body-pad");
-      this.mainpd.classList.toggle("body-pad");
-
-      this.sidebar.addEventListener("transitionend", resolve, { once: true });
-    });
-  }
-
-  toggleAccordion(btn) {
-    if (this.sidebar.classList.contains("sidebar--collapsed")) {
-      this.toggleSidebar().then(() => {
-        this.expandAccordionContent(btn);
+    let ui = this;
+    // Add the collapsed class to each of the main elements
+    ui.sidebar.classList.toggle("sidebar-collapsed");
+    ui.navbar.classList.toggle("sidebar-collapsed");
+    ui.main.classList.toggle("sidebar-collapsed");
+    // Select all of the accordion buttons
+    const accordions = document.querySelectorAll(".accordion");
+    // Check if the sidebar is collapsed
+    if (ui.sidebar.classList.contains("sidebar-collapsed")) {
+      // Close all of the sub menus
+      accordions.forEach((accordion) => {
+        accordion.classList.remove("is-open");
+        let content = accordion.nextElementSibling;
+        content.style.maxHeight = null;
       });
+      // Change the icon and text of the toggle button
+      document.querySelector("#sidebar-toggle i").className =
+        "bx bx-horizontal-right";
+      document.querySelector("#sidebar-toggle span").textContent =
+        "Expand Sidebar";
     } else {
-      this.expandAccordionContent(btn);
+      // Change the icon and text of the toggle button
+      document.querySelector("#sidebar-toggle i").className =
+        "bx bx-horizontal-left";
+      document.querySelector("#sidebar-toggle span").textContent =
+        "Collapse Sidebar";
     }
+    // Finally, dispatch a resize event to trigger plot resizing
+    window.dispatchEvent(new Event("resize"));
   }
 
-  expandAccordionContent(btn) {
+  /* Method to open/close the sidebar accordion menus */
+  toggleAccordion(btn) {
+    let ui = this;
+    // Toggle the class of the accordion button
     btn.classList.toggle("is-open");
+    // Open the side if it is closed
+    if (ui.sidebar.classList.contains("sidebar-collapsed")) {
+      ui.toggleSidebar();
+    }
+    // Open the accordion content if it is closed
     let content = btn.nextElementSibling;
-
[…]
