### 2024-04-10 20bafc3
Fix padding to be the same as the navbar
