+++ assets/v0.8e9462c5.js
@@ -4973,4 +4973,4 @@ To view the analysis pipeline and raw data, checkout the [GitHub repository](htt
           fingers on your tackpad or mouse. Change the orientation by <em>clicking and dragging</em> 
           the structure. Reset the initial zoom and position by pressing the <kbd>r</kbd> key.
         </p>
-        </details>`,"instruct"),Vn=new Gpe(gx(e),Tpe),Lpe(),Ppe(),Upe(),Dpe(),kpe()}).then(()=>{window.dispatchEvent(new Event("resize"))}),Dq(3e3)]).then(()=>{document.getElementById("loadingScreen").style.display="none"}).catch(e=>{document.getElementById("loadingScreen").style.display="none",Ms.showAlert(e.message)});async function Ipe(){const e=new URLSearchParams(window.location.search),t=e.get("data"),n=e.get("markdown");let i=zS;if(t)if(i=await r2(t)||zS,document.getElementById("remote-file").click(function(r){r.stopPropagation()}),document.getElementById("url-json-file").value=t,document.getElementById("url-markdown-file").disabled=!1,n){const r=await r2(n,"markdown");md(r),document.getElementById("url-markdown-file").value=n,i.markdown_description&&Ms.showAlert("The remote markdown description is being used instead of the one in the JSON file.","warning")}else i.markdown_description?md(i.markdown_description):yx();else md(i.markdown_description);return i}async function r2(e,t="json"){let n=null;try{const i=await fetch(e);if(!i.ok)throw new Error(`HTTP error! status: ${i.status}`);switch(t){case"json":n=await i.json();break;case"markdown":n=await i.text();break;default:throw new Error(`Unsupported format: ${t}`)}}catch(i){Ms.showAlert(`Fetch operation failed: ${i.message}`),console.error(`Fetch operation failed: ${i.message}`)}return n}function md(e){try{document.getElementById("markdown").style.display="block",document.getElementById("markdown-container").innerHTML=Zn.parse(e,{renderer:Fpe})}catch{Ms.showAlert("Markdown rendering failed. Please check that the markdown is properly formatted.")}}function yx(){document.getElementById("markdown").style.display="none",document.getElementById("markdown-container").innerHTML=""}function gx(e){const t=JSON.parse(JSON.stringify(e));return delete t.markdown_description,t}function Lpe(){_t("#local-json-file").on("change",function(){const e=document.getElementById("local-json-file");if(e.files.length===0){Ms.showAlert("Please select a JSON file to upload.");return}const t=e.files[0];if(t.type!=="application/json"){Ms.showAlert("Please select a valid JSON file.");return}const n=new FileReader;n.onload=function(){try{const i=JSON.parse(n.result);Vn.data=gx(i),Vn.initTool(),window.history.replaceState({},"",`${location.pathname}`),document.getElementById("url-json-file").value="",document.getElementById("url-markdown-file").value="",document.getElementById("url-markdown-file").disabled=!0,i.markdown_description?md(i.markdown_description):yx(),window.dispatchEvent(new Event("resize"))}catch(i){Ms.showAlert(i.message)}},n.readAsText(t)}),_t("#url-json-file").on("keyup",async function(e){if(e.key!=="Enter")return;if(!this.value){Ms.showAlert("Please enter a URL.");return}try{new URL(this.value)}catch{Ms.showAlert("Please enter a valid URL.");return}const t=await r2(this.value);document.getElementById("url-markdown-file").disabled=!1;const n=new URLSearchParams;n.set("data",this.value),window.history.replaceState({},"",`${location.pathname}?${n}`),Vn.data=
[…]
