### 2024-04-11 b297737
Harmonize pad, marg, and width between main and nav
