### 2024-04-11 b578c2c
Fix issue with menu height when data changes
