### 2024-04-11 8034eb0
Point to docs in contributing guide
