### 2024-05-08 8774494
Remove log statement
