### 2024-05-14 dms-viz/dms-viz.github.io#85 @afrubin
Color for wild-type score
The tool assumes that 0 corresponds to wild-type activity, but for many DMS scoring methods other values are chosen (such as 1). This means the color scheme doesn't make much sense for these datasets.

It would be nice if the wild-type activity score were an option that the user can set, and would rescale the heatmap.

Here's an example of one such score from [PMID 29785012](https://www.ncbi.nlm.nih.gov/pubmed/29785012) (also available from [MaveDB](https://mavedb.org/#/score-sets/urn:mavedb:00000013-a-1)).

A minimal JSON file for dms-viz is attached:
[pten_dms_viz.json](https://github.com/dms-viz/dms-viz.github.io/files/15303052/pten_dms_viz.json)

This issue is part of https://github.com/openjournals/joss-reviews/issues/6129
