### 2024-05-14 #6129 @afrubin
@WillHannon-MCB the changes to the code all look great, thanks. However it doesn't look like the paper has been updated in response to my [previous comments](https://github.com/openjournals/joss-reviews/issues/6129#issuecomment-1923210810) in the thread.

I also added one new issue (https://github.com/dms-viz/dms-viz.github.io/issues/85) that came up with a dataset I tried to visualize. That issue includes a copy of the .json file dms-viz is expecting with the scores, so you should feel free to add that to the examples if you want a non-viral one.
