### 2024-05-15 #6129 @WillHannon-MCB
@afrubin, can you recommend a non-viral dataset and accompanying paper so I can add it as an example? I don't have a good sense of the available datasets.
