+++ CHANGELOG.md
@@ -479,7 +479,7 @@
 
 - N/A
 
-## [0.6.4] - 2024-05-15
+## [0.6.5] - 2024-05-15
 
 ### Added
 
