+++ paper/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 # Summary and Purpose
 
-Understanding how mutations impact a protein's functions is valuable for many types of biological questions. High-throughput techniques such as deep-mutational scanning (DMS) have greatly expanded the number of mutation-function datasets. For instance, DMS has been used to determine how mutations to viral proteins affect antibody escape [@dadonaitePseudovirusSystemEnables2023], receptor affinity [@starrDeepMutationalScanning2020], and essential functions such as viral genome transcription and replication [@liDeepMutationalScanning2023]. With the growth of sequence databases, in some cases the effects of mutations can also be inferred from phylogenies of natural sequences [@bloomFitnessEffectsMutations2023] (\autoref{fig:figure1}).
+Understanding how mutations impact a protein's functions is valuable for addressing many biological questions. Deep-mutational scanning (DMS) offers an approach to characterize the impact of a huge number of mutations in parallel [@fowlerDeepMutationalScanning2014]. The wide application of DMS has greatly expanded the number of mutation-function datasets [@fowlerAtlasVariantEffects2023]. For instance, DMS has been used to determine how mutations to viral proteins affect antibody escape [@dadonaitePseudovirusSystemEnables2023], receptor affinity [@starrDeepMutationalScanning2020], and essential functions such as viral genome transcription and replication [@liDeepMutationalScanning2023]. With the growth of sequence databases, in some cases the effects of mutations can also be inferred from phylogenies of natural sequences [@bloomFitnessEffectsMutations2023] (\autoref{fig:figure1}).
 
 The mutation-based data generated by these approaches is often best understood in the context of a protein’s 3D structure; for instance, to assess questions like how mutations that affect antibody escape relate to the physical antibody binding epitope on the protein. However, current approaches for visualizing mutation data in the context of a protein’s structure are often cumbersome and require multiple steps and softwares. To streamline the visualization of mutation-associated data in the context of a protein structure, we developed a web-based tool, `dms-viz`. With `dms-viz`, users can straightforwardly visualize mutation-based data such as those from DMS experiments in the context of a 3D protein model in an interactive format. See <https://dms-viz.github.io/> to use `dms-viz`.
 
@@ -93,9 +93,15 @@ The influenza RNA-dependent RNA polymerase (RdRp) is essential to viral replicat
 
 See how `dms-viz` can provide this dataset as an interactive resource [here](https://dms-viz.github.io/v0/?data=https%3A%2F%2Fraw.githubusercontent.com%2Fdms-viz%2Fconfigure_dms_viz%2Fmain%2Ftests%2FIAV-PB1-DMS%2Foutput%2FIAV-PB1-DMS.json&n=true&s=mean).
 
+## 4. Visualizing the pathogenicity of genetic variants of an important tumor suppressor
+
+PTEN is a tumor suppressor that is inactivated in many cancers. Matreyek and Starita used variant abundance by massively parallel sequencing (VAMP-seq) to characterize the effect of thousands of mutations on the intracellular abundance of PTEN [@matreyekMultiplexAssessmentProtein2018].
+
+See how `dms-viz` can be used to identify clinically relevant mutations in human proteins [here]().
+
 # Conclusion
 
-We designed `dms-viz`  as a practical and user-friendly approach to visualizing mutation-associated data in the context of protein structures.  Because `dms-viz` is capable of handling various data types and has options for both sharing and privacy, it should be applicable to visualization of a wide range of datasets.
[…]
