+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{bloomFitnessEffectsMutations2023,
-  title = {Fitness Effects of~Mutations to {{SARS-CoV-2}} Proteins},
+  title = {Fitness Effects of Mutations to {{SARS-CoV-2}} Proteins},
   author = {Bloom, Jesse D. and Neher, Richard A.},
   year = {2023},
   journal = {Virus Evolution},
@@ -17,10 +17,10 @@
 
 @misc{bobyOpenScienceDiscovery2023,
   title = {Open {{Science Discovery}} of {{Potent Non-Covalent SARS-CoV-2 Main Protease Inhibitors}}},
-  author = {Boby, Melissa L. and Fearon, Daren and Ferla, Matteo and Filep, Mihajlo and Koekemoer, Lizb{\'e} and Robinson, Matthew C. and Consortium, The COVID Moonshot and Chodera, John D. and Lee, Alpha A. and London, Nir and von Delft, Annette and von Delft, Frank},
+  author = {Boby, Melissa L. and Fearon, Daren and Ferla, Matteo and Filep, Mihajlo and Koekemoer, Lizb{\'e} and Robinson, Matthew C. and Consortium, The COVID Moonshot and Chodera, John D. and Lee, Alpha A. and London, Nir and {von Delft}, Annette and {von Delft}, Frank},
   year = {2023},
   month = sep,
-  primaryclass = {New Results},
+  eprintclass = {New Results},
   pages = {2020.10.29.339317},
   publisher = {{bioRxiv}},
   doi = {10.1101/2020.10.29.339317},
@@ -28,9 +28,9 @@
   abstract = {We report the results of the COVID Moonshot, a fully open-science, crowd sourced, structure-enabled drug discovery campaign targeting the SARS-CoV-2 main protease. We discovered a non-covalent, non-peptidic inhibitor scaffold with lead-like properties that is differentiated from current main protease inhibitors. Our approach leveraged crowdsourcing, machine learning, exascale molecular simulations, and high-throughput structural biology and chemistry. We generated a detailed map of the structural plasticity of the SARS-CoV-2 main protease, extensive structure-activity relationships for multiple chemotypes, and a wealth of biochemical activity data. All compound designs ({$>$}18,000 designs), crystallographic data ({$>$}840 ligand-bound X-ray structures), assay data ({$>$}10,000 measurements), and synthesized molecules ({$>$}2,400 compounds) for this campaign were shared rapidly and openly, creating a rich open and IP-free knowledgebase for future anti-coronavirus drug discovery.},
   archiveprefix = {bioRxiv},
   chapter = {New Results},
-  copyright = {\textcopyright{} 2023, Posted by Cold Spring Harbor Laboratory. This pre-print is available under a Creative Commons License (Attribution 4.0 International), CC BY 4.0, as described at http://creativecommons.org/licenses/by/4.0/},
+  copyright = {{\textcopyright} 2023, Posted by Cold Spring Harbor Laboratory. This pre-print is available under a Creative Commons License (Attribution 4.0 International), CC BY 4.0, as described at http://creativecommons.org/licenses/by/4.0/},
   langid = {english},
-  file = {/Users/will/Zotero/storage/ZWLEJESD/Boby et al. - 2023 - Open Science Discovery of Potent Non-Covalent SARS.pdf}
+  file = {/Users/will/Zotero/storage/EAESIVEC/Boby et al. - 2023 - Open Science Discovery of Potent Non-Covalent SARS.pdf}
 }
 
 @article{dadonaitePseudovirusSystemEnables2023,
@@ -49,7 +49,64 @@
   pmcid = {PMC9922669},
   pmid = {36868218},
[…]
