### 2024-05-16 4494ee6
Allow for non-symmetric color scale
