### 2024-05-17 d226872
Only show the min and max values of the heatmap scale
