### 2024-05-17 52fa000
Clear URI state before initation
