+++ CHANGELOG.md
@@ -502,10 +502,11 @@
 ### Added
 
 - Allow users to set the min, max, and center of the color scales.
+- Allow for color scales that aren't symmetric around 0.
 
 ### Changed
 
-- N/A
+- Clear the URI parameters when data is loaded from a local file.
 
 ### Deprecated
 
