### 2024-05-17 dms-viz/dms-viz.github.io#85 @WillHannon-MCB
You can now set the limits and center of the heatmap's color scale. Note that this doesn't change how the protein's color scale is calculated.
