### 2024-05-17 #6129 @WillHannon-MCB
Hi @afrubin, I've updated the [manuscript](https://github.com/dms-viz/dms-viz.github.io/tree/submit-joss-manuscript/paper) with your comments. The only thing missing is the link to a human example. I'm planning to use the PTEN VAMPseq dataset you sent. I've got a [minimal example](https://dms-viz.github.io/v0/?data=https%3A%2F%2Fraw.githubusercontent.com%2Fdms-viz%2Fconfigure_dms_viz%2Fmain%2Ftests%2FPTEN-VAMPseq%2Foutput%2FPTEN-VAMPseq.json), but the scale limits are currently arbitrary. Would you mind sharing a `configure-dms-viz` command to make a meaningful visualization with this dataset?

I've also made it possible to set the center of the heatmap's scale. The updated [documentation](https://dms-viz.github.io/dms-viz-docs/preparing-data/command-line-api/#heatmap-limits) has the implementation details.
