### 2024-05-20 6b67b7d
Let users set the default summary and floor
