### 2024-05-22 0ff5ebb
Allow summary and floor to be set by the dataset config
