### 2024-05-25 #6129 @csoneson
👋🏻 Checking in here - @jkanche, did you have a chance to take a look at the updated submission?
@afrubin - could you also give an update on where things are at from your perspective (it looks like the issues that were opened above have been closed, but most items in your checklist are unticked, so I just want to make sure I understand where we're standing 🙂)?
Thanks!
