+++ assets/v0.65ff54cf.js
@@ -4973,4 +4973,4 @@ To view the analysis pipeline and raw data, checkout the [GitHub repository](htt
           fingers on your tackpad or mouse. Change the orientation by <em>clicking and dragging</em> 
           the structure. Reset the initial zoom and position by pressing the <kbd>r</kbd> key.
         </p>
-        </details>`,"instruct"),Vn=new Gpe(bx(e),KE),Ipe(),Lpe(),Ppe(),Upe(),Dpe()}).then(()=>{if(window.dispatchEvent(new Event("resize")),!KE.sidebar.classList.contains("sidebar-collapsed")){const e=document.getElementById("chart-btn");KE.toggleAccordion(e)}}),Dq(3e3)]).then(()=>{document.getElementById("loadingScreen").style.display="none"}).catch(e=>{document.getElementById("loadingScreen").style.display="none",Ms.showAlert(e.message)});async function Fpe(){const e=new URLSearchParams(window.location.search),t=e.get("data"),n=e.get("markdown");let i=VN;if(t)if(i=await i2(t)||VN,document.getElementById("remote-file").click(function(r){r.stopPropagation()}),document.getElementById("url-json-file").value=t,document.getElementById("url-markdown-file").disabled=!1,n){const r=await i2(n,"markdown");md(r),document.getElementById("url-markdown-file").value=n,i.markdown_description&&Ms.showAlert("The remote markdown description is being used instead of the one in the JSON file.","warning")}else i.markdown_description?md(i.markdown_description):gx();else md(i.markdown_description);return i}async function i2(e,t="json"){let n=null;try{const i=await fetch(e);if(!i.ok)throw new Error(`HTTP error! status: ${i.status}`);switch(t){case"json":n=await i.json();break;case"markdown":n=await i.text();break;default:throw new Error(`Unsupported format: ${t}`)}}catch(i){Ms.showAlert(`Fetch operation failed: ${i.message}`),console.error(`Fetch operation failed: ${i.message}`)}return n}function md(e){try{document.getElementById("markdown").style.display="block",document.getElementById("markdown-container").innerHTML=Zn.parse(e,{renderer:Tpe})}catch{Ms.showAlert("Markdown rendering failed. Please check that the markdown is properly formatted.")}}function gx(){document.getElementById("markdown").style.display="none",document.getElementById("markdown-container").innerHTML=""}function bx(e){const t=JSON.parse(JSON.stringify(e));return delete t.markdown_description,t}function Ipe(){Et("#local-json-file").on("change",function(){const e=document.getElementById("local-json-file");if(e.files.length===0){Ms.showAlert("Please select a JSON file to upload.");return}const t=e.files[0];if(t.type!=="application/json"){Ms.showAlert("Please select a valid JSON file.");return}const n=new FileReader;n.onload=function(){try{const r=JSON.parse(n.result);Vn.data=bx(r),window.history.replaceState({},"",`${location.pathname}`),document.getElementById("url-json-file").value="",document.getElementById("url-markdown-file").value="",document.getElementById("url-markdown-file").disabled=!0,Vn.initTool(),r.markdown_description?md(r.markdown_description):gx();var i=document.querySelectorAll(".submenu");i.forEach(function(a){var s=a.querySelector(".accordion"),o=a.querySelector(".accordion-content");s&&o&&s.classList.contains("is-open")&&(o.style.maxHeight=o.scrollHeight+"px")}),window.dispatchEvent(new Event("resize"))}catch(r){Ms.showAlert(r.message)}},n.readAsText(t)}),Et("#url-json-file").on("keyup",async function(e){if(e.
[…]
