### 2024-05-29 10704a2
Add comment about JSON schema
