### 2024-06-05 #6129 @afrubin
@WillHannon-MCB Looks great, sorry for the delay in getting back to this! One very minor thing is that the link in the paper for the PTEN dataset doesn't seem to work. Here is an updated file that uses the `--heatmap-limits` option to set the midpoint of the heatmap to 0. It would obviously be better if the colors in the heatmap matched the protein colors, but I don't want to delay things further.
[pten_example.json](https://github.com/user-attachments/files/15572266/pten_example.json)

@csoneson Aside from the above, I think this is ready to progress to the next steps. Thank you for your kind reminders and your patience with my busy travel schedule!
