### 2024-06-13 eed6880
Sync protein and chart value for floor
