+++ assets/v0.5d78c884.js
@@ -4901,7 +4901,7 @@ vClipCenter = -( modelViewMatrix * vec4( clipCenter, 1.0 ) ).xyz;
 `),s=0,o=0;n.useColor=!1;for(var c=0;c<a.length;c++){var A=a[c];if((A=A.trim())!==""){o>=i.elements[s].count&&(s++,o=0);var l=this.parseASCIIElement(i.elements[s].properties,A);this.handleElement(n,i.elements[s].name,l),o++}}return this.postProcess(n)},postProcess:function(e){if(e.useColor){for(var t=0;t<e.faces.length;t++)e.faces[t].vertexColors=[e.colors[e.faces[t].a],e.colors[e.faces[t].b],e.colors[e.faces[t].c]];e.elementsNeedUpdate=!0}return e.computeBoundingSphere(),e},handleElement:function(e,t,n){if(t==="vertex"){if(e.vertices.push(new P(n.x,n.y,n.z)),"red"in n&&"green"in n&&"blue"in n){e.useColor=!0;var i=new pt;i.setRGB(n.red/255,n.green/255,n.blue/255),e.colors.push(i)}}else if(t==="face"){var r=n.vertex_indices;r.length===3?e.faces.push(new hl(r[0],r[1],r[2])):r.length===4&&e.faces.push(new hl(r[0],r[1],r[3]),new hl(r[1],r[2],r[3]))}},binaryRead:function(e,t,n,i){switch(n){case"int8":case"char":return[e.getInt8(t),1];case"uint8":case"uchar":return[e.getUint8(t),1];case"int16":case"short":return[e.getInt16(t,i),2];case"uint16":case"ushort":return[e.getUint16(t,i),2];case"int32":case"int":return[e.getInt32(t,i),4];case"uint32":case"uint":return[e.getUint32(t,i),4];case"float32":case"float":return[e.getFloat32(t,i),4];case"float64":case"double":return[e.getFloat64(t,i),8]}},binaryReadElement:function(e,t,n,i){for(var r,a={},s=0,o=0;o<n.length;o++)if(n[o].type==="list"){var c=[],A=(r=this.binaryRead(e,t+s,n[o].countType,i))[0];s+=r[1];for(var l=0;l<A;l++)r=this.binaryRead(e,t+s,n[o].itemType,i),c.push(r[0]),s+=r[1];a[n[o].name]=c}else r=this.binaryRead(e,t+s,n[o].type,i),a[n[o].name]=r[0],s+=r[1];return[a,s]},parseBinary:function(e){for(var t,n=new an,i=this.parseHeader(this.bin2str(e)),r=i.format==="binary_little_endian",a=new DataView(e,i.headerLength),s=0,o=0;o<i.elements.length;o++)for(var c=0;c<i.elements[o].count;c++){s+=(t=this.binaryReadElement(a,s,i.elements[o].properties,r))[1];var A=t[0];this.handleElement(n,i.elements[o].name,A)}return this.postProcess(n)}};var $de=function(e){function t(){e.apply(this,arguments)}e&&(t.__proto__=e),t.prototype=Object.create(e&&e.prototype),t.prototype.constructor=t;var n={type:{configurable:!0}};return n.type.get=function(){return"ply"},t.prototype.getLoader=function(){return new mx},Object.defineProperties(t.prototype,n),t}(AL);Jt.add("ply",$de);var jde=function(e){function t(i,r){var a=r||{};e.call(th
[…]
