### 2024-06-15 #6129 @csoneson
Hi @WillHannon-MCB - I have gone through the submission and have a few small suggestions for the paper:

* Could you please add the country (spelled out) to the affiliations?
* The sentence on line 27 sounds a bit strange to me - perhaps replace 'See' with 'Visit'?
* In Boby et al, could you please fix the author 'Consortium, T. C. M.' so that it's spelled out correctly?

I think it would also be preferable to have the figures placed in the paper close to where they are mentioned.

Finally, note that your paper is quite a bit longer than the JOSS [guidelines](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) (250-1000 words). Could you please look into shortening it a bit (e.g. by moving some of the descriptions of the case studies to an online location)?

You can then generate a new proof with `@editorialbot generate pdf`.

Thanks!
