### 2024-06-17 #6129 @WillHannon-MCB
@csoneson, the citation for Boby et al looks fine when I compile the LaTeX into a PDF. Does that suggest that the bibtex is fine?

![Screenshot 2024-06-17 at 2 35 50 PM](https://github.com/openjournals/joss-reviews/assets/62000409/022dbfdb-9882-4028-84e9-1c7c8bcdd90b)
