+++ paper/paper.md
@@ -15,13 +15,13 @@ authors:
     orcid: 0000-0003-1267-3408
     affiliation: "2, 3, 4" 
 affiliations:
- - name: Molecular and Cellular Biology Graduate Program, University of Washington, Seattle, WA 98109
+ - name: Molecular and Cellular Biology Graduate Program, University of Washington, Seattle, Washington 98109, United States of America
    index: 1
- - name: Basic Sciences and Computational Biology, Fred Hutchinson Cancer Research Center, Seattle, WA 98109
+ - name: Basic Sciences and Computational Biology, Fred Hutchinson Cancer Research Center, Seattle, Washington 98109, United States of America
    index: 2
- - name: Department of Genome Sciences, University of Washington, Seattle, WA 98109
+ - name: Department of Genome Sciences, University of Washington, Seattle, Washington 98109, United States of America
    index: 3
- - name: Howard Hughes Medical Institute, Seattle, WA 98109
+ - name: Howard Hughes Medical Institute, Seattle, Washington 98109, United States of America
    index: 4
 date: 26 October 2023
 bibliography: paper.bib
@@ -31,7 +31,7 @@ bibliography: paper.bib
 
 Understanding how mutations impact a protein's functions is valuable for addressing many biological questions. Deep-mutational scanning (DMS) offers an approach to characterize the impact of a huge number of mutations in parallel [@fowlerDeepMutationalScanning2014]. The wide application of DMS has greatly expanded the number of mutation-function datasets [@fowlerAtlasVariantEffects2023]. For instance, DMS has been used to determine how mutations to viral proteins affect antibody escape [@dadonaitePseudovirusSystemEnables2023], receptor affinity [@starrDeepMutationalScanning2020], and essential functions such as viral genome transcription and replication [@liDeepMutationalScanning2023]. With the growth of sequence databases, in some cases the effects of mutations can also be inferred from phylogenies of natural sequences [@bloomFitnessEffectsMutations2023] (\autoref{fig:figure1}).
 
-The mutation-based data generated by these approaches is often best understood in the context of a protein’s 3D structure; for instance, to assess questions like how mutations that affect antibody escape relate to the physical antibody binding epitope on the protein. However, current approaches for visualizing mutation data in the context of a protein’s structure are often cumbersome and require multiple steps and softwares. To streamline the visualization of mutation-associated data in the context of a protein structure, we developed a web-based tool, `dms-viz`. With `dms-viz`, users can straightforwardly visualize mutation-based data such as those from DMS experiments in the context of a 3D protein model in an interactive format. See <https://dms-viz.github.io/> to use `dms-viz`.
+The mutation-based data generated by these approaches is often best understood in the context of a protein’s 3D structure; for instance, to assess questions like how mutations that affect antibody escape relate to the physical antibody binding epitope on the protein. However, current approaches for visualizing mutation data in the context of a protein’s structure are often cumbersome and require multiple steps and softwares. To streamline the visualization of mutation-associated data in the context of a protein structure, we developed a web-based tool, `dms-viz`. With `dms-viz`, users can straightforwardly visualize mutation-based data such as those from DMS experiments in the context of a 3D protein model in an interactive format. Visit <https://dms-viz.github.io/> to use `dms-viz`.
 
 # Statement of Need
 
@@ -81,7 +81,7 @@ See how `dms-viz` can be used to interactively visualize datasets with multiple 
 
[…]
