### 2024-06-18 96a8a8d
Shorten examples
