+++ paper/paper.md
@@ -71,33 +71,23 @@ To ensure the visualization remains compact, all configuration options are tucke
 
 # Examples
 
+For additional scientific background and a walkthrough of the code that generates these visualizations, visit the [documentation](https://dms-viz.github.io/dms-viz-docs/visualizing-data/vignettes/).
+
 ## 1. Mapping the neutralization profile of antibodies and sera against HIV envelope
 
-@radfordMappingNeutralizingSpecificity2023 mapped mutations to HIV envelope (Env) that affect neutralization by polyclonal human serum using a pseudotyping-based deep mutational scanning platform [@radfordMappingNeutralizingSpecificity2023]. One aim of their study was to examine how the sites of escape mutations related to HIV Env’s structure. `dms-viz` excels in generating these visualizations, especially for intra-experimental comparisons. Using `dms-viz`, it is possible to show multiple antibody footprints on a single summary plot.
-
-See how `dms-viz` can be used to interactively visualize datasets with multiple conditions [here](https://dms-viz.github.io/v0/?data=https%3A%2F%2Fraw.githubusercontent.com%2Fdms-viz%2Fconfigure_dms_viz%2Fmain%2Ftests%2FHIV-Envelope-BF520-DMS%2Foutput%2FHIV-Envelope-BF520-DMS.json&e=IDC508&ce=%255B%25221%2522%252C%25222%2522%255D).
+@radfordMappingNeutralizingSpecificity2023 mapped mutations to HIV envelope (Env) that affect neutralization by polyclonal human serum using a pseudotyping-based deep mutational scanning platform [@radfordMappingNeutralizingSpecificity2023]. See how `dms-viz` can be used to interactively visualize datasets with multiple antibody footprints on a single summary plot [here](https://dms-viz.github.io/v0/?data=https%3A%2F%2Fraw.githubusercontent.com%2Fdms-viz%2Fconfigure_dms_viz%2Fmain%2Ftests%2FHIV-Envelope-BF520-DMS%2Foutput%2FHIV-Envelope-BF520-DMS.json&e=IDC508&ce=%255B%25221%2522%252C%25222%2522%255D).
 
 ## 2. Using mutation-fitness data to augment structure-guided drug design
 
-Bloom and Neher developed a method to estimate the fitness effects of mutations to all SARS-CoV-2 proteins by analyzing millions of human SARS-CoV-2 sequences  [@bloomFitnessEffectsMutations2023]. These mutation-fitness estimates are useful for purposes such as attempting to design antiviral drugs that target functionally constrained sites where resistance is unlikely to emerge.
-
-By merging Bloom and Neher's data with structural views of a viral target like the SARS-CoV-2 main protease (Mpro) in complex with a bound ligand such as MAT-POS-e194df51-1 from the COVID Moonshot project [@bobyOpenScienceDiscovery2023], `dms-viz` offers an intuitive way to visualize whether a ligand is targeting a mutationally tolerant binding pocket. Computational chemists can incorporate this information into the design process by screening for compounds that target sites where mutations have negative effects on viral fitness.
-
[…]
