### 2024-06-18 #6129 @jkanche
Yes I am, The only thing I noticed before was a missing example for one of the usecases, but it seems to be added in the newer version of the article! looks great!
