### 2024-07-01 #6129 @csoneson
@WillHannon-MCB I was referring to the posts above:

* https://github.com/openjournals/joss-reviews/issues/6129#issuecomment-2176746107
* https://github.com/openjournals/joss-reviews/issues/6129#issuecomment-2170628928

I don't think these have been fully addressed (the consortium author still looks strange, the figures are not integrated into the text, and the length of the paper is still quite a bit over the guidelines).
