### 2024-07-02 b4eed08
Integrate figures into text
