+++ paper/paper.md
@@ -29,25 +29,27 @@ bibliography: paper.bib
 
 # Summary and Purpose
 
-Understanding how mutations impact a protein's functions is valuable for addressing many biological questions. Deep-mutational scanning (DMS) offers an approach to characterize the impact of a huge number of mutations in parallel [@fowlerDeepMutationalScanning2014]. The wide application of DMS has greatly expanded the number of mutation-function datasets [@fowlerAtlasVariantEffects2023]. For instance, DMS has been used to determine how mutations to viral proteins affect antibody escape [@dadonaitePseudovirusSystemEnables2023], receptor affinity [@starrDeepMutationalScanning2020], and essential functions such as viral genome transcription and replication [@liDeepMutationalScanning2023]. With the growth of sequence databases, in some cases the effects of mutations can also be inferred from phylogenies of natural sequences [@bloomFitnessEffectsMutations2023] (\autoref{fig:figure1}).
+Many biological questions require an understanding of how mutations impact a protein's functions. Deep-mutational scanning (DMS) offers an approach to characterize the impact of a huge number of mutations in parallel [@fowlerDeepMutationalScanning2014]. The wide application of DMS has greatly increased the number of mutation-function datasets [@fowlerAtlasVariantEffects2023]. For instance, DMS has been used to determine how mutations to viral proteins affect antibody escape [@dadonaitePseudovirusSystemEnables2023], receptor affinity [@starrDeepMutationalScanning2020], and essential functions such as viral genome transcription and replication [@liDeepMutationalScanning2023]. In some cases, the effects of mutations can also be inferred from phylogenies of natural sequences [@bloomFitnessEffectsMutations2023] (\autoref{fig:figure1}).
 
-The mutation-based data generated by these approaches is often best understood in the context of a protein’s 3D structure; for instance, to assess questions like how mutations that affect antibody escape relate to the physical antibody binding epitope on the protein. However, current approaches for visualizing mutation data in the context of a protein’s structure are often cumbersome and require multiple steps and softwares. To streamline the visualization of mutation-associated data in the context of a protein structure, we developed a web-based tool, `dms-viz`. With `dms-viz`, users can straightforwardly visualize mutation-based data such as those from DMS experiments in the context of a 3D protein model in an interactive format. Visit <https://dms-viz.github.io/> to use `dms-viz`.
+![Large mutation-associated datasets can be used to map antibody footprints, assess the impact of mutations on protein function, and identify patterns of selection from natural mutation frequencies.\label{fig:figure1}](figures/figure-1.png)
+
+The mutation-based data generated by these approaches is better understood in the context of a protein’s 3D structure. However, current approaches for visualizing mutation data in the context of a protein’s structure are often cumbersome and require multiple steps and software. To streamline the visualization of mutation-associated data in the context of a protein structure, we developed a web-based tool, `dms-viz`. With `dms-viz`, users can straightforwardly visualize mutation-based data such as those from DMS experiments in the context of a 3D protein model in an interactive format. Visit <https://dms-viz.github.io/> to use `dms-viz`.
 
 # Statement of Need
 
 We wanted `dms-viz` to provide the following functionalities:
 
-1. **Provide structural context**: The main objective of `dms-viz` is to simplify the process of visualizing mutation data with structural context by superimposing mutation measurements on a 3D protein structure. Additionally, it provides extensive control over the visual representation of the 3D structure.
[…]
