### 2024-07-02 #6129 @WillHannon-MCB
@csoneson, I integrated the figures into the text. I'll try to cut the words down more, but it will be tough without removing a whole section.
