### 2024-07-02 cd6c44a
Remove redudant example
