+++ paper/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 # Summary and Purpose
 
-Many biological questions require an understanding of how mutations impact a protein's functions. Deep-mutational scanning (DMS) offers an approach to characterize the impact of a huge number of mutations in parallel [@fowlerDeepMutationalScanning2014]. The wide application of DMS has greatly increased the number of mutation-function datasets [@fowlerAtlasVariantEffects2023]. For instance, DMS has been used to determine how mutations to viral proteins affect antibody escape [@dadonaitePseudovirusSystemEnables2023], receptor affinity [@starrDeepMutationalScanning2020], and essential functions such as viral genome transcription and replication [@liDeepMutationalScanning2023]. In some cases, the effects of mutations can also be inferred from phylogenies of natural sequences [@bloomFitnessEffectsMutations2023] (\autoref{fig:figure1}).
+Many biological questions require an understanding of how mutations impact a protein's functions. Deep mutational scanning (DMS) offers an approach to characterize the impact of a huge number of mutations in parallel [@fowlerDeepMutationalScanning2014]. The wide application of DMS has greatly increased the number of mutation-function datasets [@fowlerAtlasVariantEffects2023]. For instance, DMS has been used to determine how mutations to viral proteins affect antibody escape [@dadonaitePseudovirusSystemEnables2023], receptor affinity [@starrDeepMutationalScanning2020], and essential functions such as viral genome transcription and replication [@liDeepMutationalScanning2023]. In some cases, the effects of mutations can also be inferred from phylogenies of natural sequences [@bloomFitnessEffectsMutations2023] (\autoref{fig:figure1}).
 
 ![Large mutation-associated datasets can be used to map antibody footprints, assess the impact of mutations on protein function, and identify patterns of selection from natural mutation frequencies.\label{fig:figure1}](figures/figure-1.png)
 
