### 2024-07-17 #6129 @Kevin-Mattheus-Moerman
@WillHannon-MCB as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Everything seems in order, so I will now proceed to accept this submission.
